+++ packages/cli-tools/bin/streamr-internal-sponsorship-create.ts
@@ -9,6 +9,8 @@ import { parseEther } from 'ethers'
 interface Options extends BaseOptions {
     earningsPerSecond: string
     minOperatorCount?: number
+    maxOperatorCount?: number
+    minStakeDuration?: number
 }
 
 createClientCommand(async (client: StreamrClient, streamId: string, options: Options) => {
@@ -17,12 +19,16 @@ createClientCommand(async (client: StreamrClient, streamId: string, options: Opt
         earningsPerSecond: parseEther(options.earningsPerSecond),
         deployer: await client.getSigner(),
         minOperatorCount: options.minOperatorCount,
+        maxOperatorCount: options.maxOperatorCount,
+        minStakeDuration: options.minStakeDuration,
         environmentId: client.getConfig().environment
     })
     console.info(JSON.stringify({ address: await contract.getAddress() }, undefined, 4))
 })
     .description('create sponsorship')
     .arguments('<streamId>')
-    .requiredOption('-e, --earnings-per-second <number>', 'Earnings per second')
-    .option('-c, --min-operator-count <number>', 'Minimum operator count')
+    .requiredOption('-e, --earnings-per-second <number>', 'Earnings per second in data tokens')
+    .option('--min-operator-count <number>', 'Minimum operator count')
+    .option('--max-operator-count <number>', 'Maximum operator count')
+    .option('--min-stake-duration <number>', 'Minimum time in seconds a stake must be held before it can be unstaked without penalty')
     .parseAsync()